requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiofiles==23.2.1
